Move BlockEncoding transformations to separate folder
