+++ src/qrisp/block_encodings/transformations/__init__.py
@@ -16,24 +16,8 @@
 ********************************************************************************
 """
 
-from .block_encoding_base import BlockEncoding
-from .constructors import (
-    build_from_array,
-    build_from_eye,
-    build_from_lcu,
-    build_from_operator,
-    build_from_projector,
-)
-from .transformations import apply_inv, apply_poly, apply_sim
+from .inv import apply_inv
+from .sim import apply_sim
+from .poly import apply_poly
 
-BlockEncoding.from_array = classmethod(build_from_array)
-BlockEncoding.from_eye = classmethod(build_from_eye)
-BlockEncoding.from_lcu = classmethod(build_from_lcu)
-BlockEncoding.from_operator = classmethod(build_from_operator)
-BlockEncoding.from_projector = classmethod(build_from_projector)
-
-BlockEncoding.inv = apply_inv
-BlockEncoding.poly = apply_poly
-BlockEncoding.sim = apply_sim
-
-__all__ = ["BlockEncoding"]
+__all__ = ["apply_inv", "apply_sim", "apply_poly"]
